pyproject.toml
[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >= 3.4, < 4"]

[project]
authors = [
  {name = "Jacqueline Ryan", email = "Jacqueline.Ryan@jpl.caltech.edu"},
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Typing :: Typed",
  "Environment :: Console",
]
dependencies = [
  "gltflib >= 1.0.13",
  "netCDF4 >= 1.7.1.post1",
  "numpy",
  "pillow",
  "pynrrd",
]
description = "Utility for exporting (cloud tomography) netCDF files to other formats."
name = "ncexport"
readme = "README.md"
requires-python = ">=3.9,<3.12"
version = "0.2.0"

[project.optional-dependencies]
dev = [
  "black",
  "flit",
  "mypy",
  "pyright",
]

[tool.flit.module]
name = "ncexport"

# [options]
# py_modules = ["nc2gltf"]

[project.scripts]
nc2gltf = "ncexport.cli:main"
nc2nrrd = "ncexport.cli:main_nrrd"
ncradiance = "ncexport.cli:main_rad"

[tool.pyright]
include = ["src"]
pythonVersion = "3.11"
strict = ["src"]

[tool.mypy]
python_version = "3.11"
show_error_codes = true
